requirements.txt
bokeh==0.12.10
certifi==2017.7.27.1
click==6.7
cloudpickle==0.4.0
cycler==0.10.0
dask==0.15.4
decorator==4.1.2
distributed==1.19.1
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
heapdict==1.0.0
imageio==2.2.0
itsdangerous==0.24
Jinja2==2.9.6
locket==0.2.0
MarkupSafe==1.0
matplotlib==2.0.2
msgpack-python==0.4.8
networkx==2.0
numpy==1.12.1
olefile==0.44
pandas==0.21.0
partd==0.3.8
Pillow==4.2.1
protobuf==3.4.1
psutil==5.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.0
scipy==1.0.0
six==1.11.0
sortedcontainers==1.5.7
tblib==1.3.2
tensorflow==1.1.0
toolz==0.8.2
tornado==4.5.2
visitor==0.1.3
Werkzeug==0.12.2
zict==0.1.3
